Small fixes. Move CollectionsTest to another package.
